- Fix Classic Controller D-Pad Right overlapping with C-Buttons - Fix ROM header length - Use AESND_SetVoiceFrequencyRatio - Changes to support extra guVecNormalize arg in libogc2

+++ gc_input/controller-Classic.c
@@ -105,7 +105,7 @@ static button_t menu_combos[] = {
 
 static unsigned int getButtons(classic_ctrl_t* controller, float maxLMag, float maxRMag)
 {
-	unsigned int b = (unsigned)controller->btns;
+	unsigned int b = (unsigned short)controller->btns;
 	s8 stickX      = getStickValue(&controller->ljs, maxLMag, STICK_X, 7);
 	s8 stickY      = getStickValue(&controller->ljs, maxLMag, STICK_Y, 7);
 	s8 substickX   = getStickValue(&controller->rjs, maxRMag, STICK_X, 7);
